+++ PaperPlaneEvolution/src/main/java/util/Util.java
@@ -4,13 +4,13 @@ package util;
 
 public class Util {
 	
-	public static float cosf(float v) {
-		return (float)Math.cos(v);
+	public static double cosf(double v) {
+		return Math.cos(v);
 	}
-	public static float sinf(float v) {
-		return (float)Math.sin(v);
+	public static double sinf(double v) {
+		return Math.sin(v);
 	}
-	public static float tanf(float v) {
-		return (float)Math.tan(v);
+	public static double tanf(double v) {
+		return Math.tan(v);
 	}
 }
